BREAK: split kinematics module
* FEAT: make `determine_indices()` publicly available
